B: Threshold int 수정 + UI 배선(그리드/버튼 → BaseUnit 노드 API)
Part1 — Threshold 소실 제거:
- InspectorAlgoRecipe.Threshold double→int(기본 128, Blob 0~255). ApplyToRuntime 캐스트 제거(int↔int).

Part2 — UI 배선(타깃 페이지 → BaseUnit, docs/B_UI_WIRING_PLAN.md):
- 그리드 = 실 finder/inspector 직접 바인딩(P4 store.GetByTarget/FromDescriptor 제거).
  finder: Search/Train ROI·AcceptThreshold(Recipe)·MaxInstances(Config). inspector: InspectionRoi(Recipe)
  + Threshold(Cognex 한정). 편집→런타임 직접, 저장 시 노드 Collect.
- 노드 해석: _module.Algorithms.First(a => a.Finder/Inspector == 대상).
- SaveTarget = node.SaveSettings()+SaveRecipe("default") / LoadTarget = node.LoadSettings()+LoadRecipe("default")
  → A Apply/Collect 가 POCO↔런타임 동기화. 레시피명 상수 "default"(선택기 후속).
- 상태점 경로 통일: RecipePage SettingPath/AlgoStatus → Recipes/default/<모듈.알고>.recipe.json(노드 TargetPath 일치).
- SettingsPage ② 에디터(구 ParameterStore)는 다음 단계(ParameterStore 삭제)에서 정리 — 이번 무수정.

게이트: 빌드0(Common→Vision, 신규 경고0) / 정적(ic-helper-call)0(Designer 불변) / 모듈 왕복 PASS(편집→
저장→재로드→런타임 복원, 파일경로 일치) / verify 코어 FAIL0 / R2 레이아웃 보존(렌더 3/3).

+++ QMC.Vision/Modules/AlgorithmNode.cs
@@ -112,7 +112,7 @@ namespace QMC.Vision.Modules
             if (Recipe is InspectorAlgoRecipe r)
             {
                 if (r.InspectionRoi != null) _inspector.InspectionRoi = r.InspectionRoi.Clone();
-                if (_inspector is CognexInspector cog) cog.Threshold = (int)r.Threshold;
+                if (_inspector is CognexInspector cog) cog.Threshold = r.Threshold;   // int↔int (소실 제거)
             }
         }
